[NEST-FE-52] change admin pages to compnents
