+++ components/admin/ReportedPosts.tsx
@@ -2,12 +2,11 @@
 export const dynamic = "force-dynamic";
 
 import ReportCard from "@/components/ReportCard";
-import SideMenu from "@/components/SideMenu";
 import { useEffect, useState } from "react";
 import { ReportPostType, Report } from "@/types/PostType";
 import { get } from "@/app/lib/fetchInterceptor";
 
-export default function ReportedPostsPage() {
+export default function ReportedPostsComponent() {
   const [loading, setLoading] = useState(true);
   const [reportedPosts, setReportedPosts] = useState<Report[]>([]);
   const [posts, setPosts] = useState<ReportPostType[]>([]);
@@ -42,29 +41,22 @@ export default function ReportedPostsPage() {
   );
 
   return (
-    <div className="p-4 sm:ml-64">
-      <SideMenu admin />
-      <div className="pl-0 p-8 flex flex-col items-start">
-        <h1 className="text-2xl font-bold text-black mb-4 pb-4">
-          Reported Posts
-        </h1>
-        {postsWithReports.length === 0 && (
-          <div className="text-gray-500">No reported posts</div>
-        )}
-        <div className="flex flex-col gap-6 w-full">
-          {postsWithReports.map((post) => {
-            const associatedReports = reportedPosts.filter(
-              (report) => report.postId === post.id
-            );
-            return (
-              <ReportCard
-                key={post.id}
-                post={post}
-                reports={associatedReports}
-              />
-            );
-          })}
-        </div>
+    <div className="p-4">
+      <h1 className="text-2xl font-bold text-black mb-4 pb-4">
+        Reported Posts
+      </h1>
+      {postsWithReports.length === 0 && (
+        <div className="text-gray-500">No reported posts</div>
+      )}
+      <div className="flex flex-col gap-6 w-full">
+        {postsWithReports.map((post) => {
+          const associatedReports = reportedPosts.filter(
+            (report) => report.postId === post.id
+          );
+          return (
+            <ReportCard key={post.id} post={post} reports={associatedReports} />
+          );
+        })}
       </div>
     </div>
   );
