+++ components/admin/TagManagement.tsx
@@ -2,7 +2,6 @@
 export const dynamic = "force-dynamic";
 
 import Button from "@/components/Button";
-import SideMenu from "@/components/SideMenu";
 import PopupWindow from "@/components/PopupWindow";
 import { useEffect, useState } from "react";
 import { del, get, post, put } from "@/app/lib/fetchInterceptor";
@@ -18,7 +17,7 @@ interface Tag {
   name: string;
 }
 
-export default function TagManagementPage() {
+export default function TagManagementComponent() {
   const [modalIndex, setModalIndex] = useState<number | null>(null);
   const [topics, setTopics] = useState<Topic[]>([]);
   const [tags, setTags] = useState<Tag[]>([]);
@@ -259,173 +258,169 @@ export default function TagManagementPage() {
   }
 
   return (
-    <div className="p-4 sm:ml-64 bg-white">
-      <SideMenu admin />
-      {/* topic section */}
-      <div>
-        <h2 className="text-xl font-semibold mb-4 text-gray-700 mt-5">
-          Topics Management
-        </h2>
-        <div className="mb-6">
-          {topics.map((topic, index) => (
-            <div
-              key={topic.id}
-              className="flex items-center justify-between border-b border-gray-300 py-2"
-            >
-              <div className="flex flex-col">
-                <div className="flex">
-                  <span className="mr-4 text-gray-700">{index + 1}.</span>
-                  <span className="text-gray-700">{topic.name}</span>
-                </div>
-                <p className="text-gray-400 ml-8">{topic.description}</p>
+    <div className="p-4">
+      <h2 className="text-xl font-semibold mb-4 text-gray-700 mt-5">
+        Topics Management
+      </h2>
+      <div className="mb-6">
+        {topics.map((topic, index) => (
+          <div
+            key={topic.id}
+            className="flex items-center justify-between border-b border-gray-300 py-2"
+          >
+            <div className="flex flex-col">
+              <div className="flex">
+                <span className="mr-4 text-gray-700">{index + 1}.</span>
+                <span className="text-gray-700">{topic.name}</span>
               </div>
-              <Button
-                className="text-secondary underline"
-                onClick={() => showModal(1, topic.id)}
-                label="Edit"
+              <p className="text-gray-400 ml-8">{topic.description}</p>
+            </div>
+            <Button
+              className="text-secondary underline"
+              onClick={() => showModal(1, topic.id)}
+              label="Edit"
+            />
+          </div>
+        ))}
+      </div>
+      <div className="mt-6 border-gray-200 flex justify-end pt-4">
+        <Button
+          label="Add Topic"
+          onClick={() => showModal(2)}
+          data-modal-target="static-modal"
+          data-modal-toggle="static-modal"
+          className="text-secondary bg-container hover:bg-amber-100 font-bold rounded-md text-sm px-5 py-2.5"
+        />
+      </div>
+
+      {/* tag section */}
+      <h2 className="text-xl font-semibold mb-4 text-gray-700">
+        Tags Management
+      </h2>
+      <div className="flex flex-wrap">
+        {tags.map((tag: Tag) => (
+          <div
+            key={tag.id}
+            className="flex border border-gray-400 items-center bg-white text-white px-2 py-2 rounded-md mb-2 mx-2"
+          >
+            <button
+              className="text-gray-500 flex"
+              onClick={() => showModal(3, tag.id)}
+            >
+              {tag.name}
+              <svg
+                className="ml-2 w-5 h-5 text-gray-500 hover:text-tertiary"
+                aria-hidden="true"
+                xmlns="http://www.w3.org/2000/svg"
+                width="24"
+                height="24"
+                fill="none"
+                viewBox="0 0 24 24"
+              >
+                <path
+                  stroke="currentColor"
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  strokeWidth="2"
+                  d="m14.304 4.844 2.852 2.852M7 7H4a1 1 0 0 0-1 1v10a1 1 0 0 0 1 1h11a1 1 0 0 0 1-1v-4.5m2.409-9.91a2.017 2.017 0 0 1 0 2.853l-6.844 6.844L8 14l.713-3.565 6.844-6.844a2.015 2.015 0 0 1 2.852 0Z"
+                />
+              </svg>
+            </button>
+          </div>
+        ))}
+      </div>
+
+      <div className="mt-6 border-gray-200 flex justify-end pt-4 border-t border-gray-300">
+        <Button
+          label="Add Tag"
+          onClick={() => showModal(4)}
+          className="text-tertiary bg-primary hover:bg-red-200 font-bold rounded-md text-sm px-5 py-2.5"
+        />
+        {/* edit topics popup window modal */}
+        {modalIndex === 1 && (
+          <div className="fixed inset-0 bg-black bg-opacity-50 flex justify-center items-center z-50">
+            <div className="absolute inset-0 flex justify-center items-center">
+              <PopupWindow
+                title="Edit Topic"
+                titleValue={selectedTopic?.name || ""}
+                onTitleChange={(e) => handleInputChange("name", "topic", e)}
+                submitButtonText="Save"
+                onDescriptionChange={(e) =>
+                  handleInputChange("description", "topic", e)
+                }
+                descriptionValue={selectedTopic?.description || ""}
+                deleteButton={true}
+                descriptionInput={true}
+                onSubmit={() => editTopic()}
+                onClose={closeModal}
+                onDelete={() => deleteSelectedTopic()}
               />
             </div>
-          ))}
-        </div>
-        <div className="mt-6 border-gray-200 flex justify-end pt-4">
-          <Button
-            label="Add Topic"
-            onClick={() => showModal(2)}
-            data-modal-target="static-modal"
-            data-modal-toggle="static-modal"
-            className="text-secondary bg-container hover:bg-amber-100 font-bold rounded-md text-sm px-5 py-2.5"
-          />
-        </div>
+          </div>
+        )}
 
-        {/* tag section */}
-        <h2 className="text-xl font-semibold mb-4 text-gray-700">
-          Tags Management
-        </h2>
-        <div className="flex flex-wrap">
-          {tags.map((tag: Tag) => (
-            <div
-              key={tag.id}
-              className="flex border border-gray-400 items-center bg-white text-white px-2 py-2 rounded-md mb-2 mx-2"
-            >
-              <button
-                className="text-gray-500 flex"
-                onClick={() => showModal(3, tag.id)}
-              >
-                {tag.name}
-                <svg
-                  className="ml-2 w-5 h-5 text-gray-500 hover:text-tertiary"
-                  aria-hidden="true"
-                  xmlns="http://www.w3.org/2000/svg"
-                  width="24"
-                  height="24"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                >
-                  <path
-                    stroke="currentColor"
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth="2"
-                    d="m14.304 4.844 2.852 2.852M7 7H4a1 1 0 0 0-1 1v10a1 1 0 0 0 1 1h11a1 1 0 0 0 1-1v-4.5m2.409-9.91a2.017 2.017 0 0 1 0 2.853l-6.844 6.844L8 14l.713-3.565 6.844-6.844a2.015 2.015 0 0 1 2.852 0Z"
-                  />
-                </svg>
-              </button>
+        {/* add topics popup window modal */}
+        {modalIndex === 2 && (
+          <div className="fixed inset-0 bg-black bg-opacity-50 flex justify-center items-center z-50">
+            <div className="absolute inset-0 flex justify-center items-center">
+              <PopupWindow
+                title="Create a New Topic"
+                titleValue={selectedTopic?.name || ""}
+                onTitleChange={(e) => handleInputChange("name", "topic", e)}
+                submitButtonText="Add"
+                onDescriptionChange={(e) =>
+                  handleInputChange("description", "topic", e)
+                }
+                descriptionValue={selectedTopic?.description || ""}
+                descriptionInput={true}
+                deleteButton={false}
+                onSubmit={() => addTopic()}
+                onClose={closeModal}
+                onDelete={() => {}}
+              />
             </div>
-          ))}
-        </div>
+          </div>
+        )}
 
-        <div className="mt-6 border-gray-200 flex justify-end pt-4 border-t border-gray-300">
-          <Button
-            label="Add Tag"
-            onClick={() => showModal(4)}
-            className="text-tertiary bg-primary hover:bg-red-200 font-bold rounded-md text-sm px-5 py-2.5"
-          />
-          {/* edit topics popup window modal */}
-          {modalIndex === 1 && (
-            <div className="fixed inset-0 bg-black bg-opacity-50 flex justify-center items-center z-50">
-              <div className="absolute inset-0 flex justify-center items-center">
-                <PopupWindow
-                  title="Edit Topic"
-                  titleValue={selectedTopic?.name || ""}
-                  onTitleChange={(e) => handleInputChange("name", "topic", e)}
-                  submitButtonText="Save"
-                  onDescriptionChange={(e) =>
-                    handleInputChange("description", "topic", e)
-                  }
-                  descriptionValue={selectedTopic?.description || ""}
-                  deleteButton={true}
-                  descriptionInput={true}
-                  onSubmit={() => editTopic()}
-                  onClose={closeModal}
-                  onDelete={() => deleteSelectedTopic()}
-                />
-              </div>
+        {/* edit tag popup window modal */}
+        {modalIndex === 3 && (
+          <div className="fixed inset-0 bg-black bg-opacity-50 flex justify-center items-center z-50">
+            <div className="absolute inset-0 flex justify-center items-center">
+              <PopupWindow
+                title="Edit Tag"
+                titleValue={selectedTag?.name || ""}
+                onTitleChange={(e) => handleInputChange("name", "tag", e)}
+                submitButtonText="Save"
+                descriptionValue=""
+                descriptionInput={false}
+                deleteButton={true}
+                onSubmit={() => editTag()}
+                onClose={closeModal}
+                onDelete={() => deleteSelectedTag()}
+              />
             </div>
-          )}
+          </div>
+        )}
 
-          {/* add topics popup window modal */}
-          {modalIndex === 2 && (
-            <div className="fixed inset-0 bg-black bg-opacity-50 flex justify-center items-center z-50">
-              <div className="absolute inset-0 flex justify-center items-center">
-                <PopupWindow
-                  title="Create a New Topic"
-                  titleValue={selectedTopic?.name || ""}
-                  onTitleChange={(e) => handleInputChange("name", "topic", e)}
-                  submitButtonText="Add"
-                  onDescriptionChange={(e) =>
-                    handleInputChange("description", "topic", e)
-                  }
-                  descriptionValue={selectedTopic?.description || ""}
-                  descriptionInput={true}
-                  deleteButton={false}
-                  onSubmit={() => addTopic()}
-                  onClose={closeModal}
-                  onDelete={() => {}}
-                />
-              </div>
+        {/* add tag popup window modal */}
+        {modalIndex === 4 && (
+          <div className="fixed inset-0 bg-black bg-opacity-50 flex justify-center items-center z-50">
+            <div className="absolute inset-0 flex justify-center items-center">
+              <PopupWindow
+                title="Create a New Tag"
+                titleValue={selectedTag?.name || ""}
+                onTitleChange={(e) => handleInputChange("name", "tag", e)}
+                submitButtonText="Add"
+                descriptionValue=""
+                descriptionInput={false}
+                deleteButton={false}
+                onSubmit={() => addTag()}
+                onClose={closeModal}
+                onDelete={() => {}}
+              />
             </div>
-          )}
-
-          {/* edit tag popup window modal */}
-          {modalIndex === 3 && (
-            <div className="fixed inset-0 bg-black bg-opacity-50 flex justify-center items-center z-50">
-              <div className="absolute inset-0 flex justify-center items-center">
-                <PopupWindow
-                  title="Edit Tag"
-                  titleValue={selectedTag?.name || ""}
-                  onTitleChange={(e) => handleInputChange("name", "tag", e)}
-                  submitButtonText="Save"
-                  descriptionValue=""
-                  descriptionInput={false}
-                  deleteButton={true}
-                  onSubmit={() => editTag()}
-                  onClose={closeModal}
-                  onDelete={() => deleteSelectedTag()}
-                />
-              </div>
-            </div>
-          )}
-
-          {/* add tag popup window modal */}
-          {modalIndex === 4 && (
-            <div className="fixed inset-0 bg-black bg-opacity-50 flex justify-center items-center z-50">
-              <div className="absolute inset-0 flex justify-center items-center">
-                <PopupWindow
-                  title="Create a New Tag"
-                  titleValue={selectedTag?.name || ""}
-                  onTitleChange={(e) => handleInputChange("name", "tag", e)}
-                  submitButtonText="Add"
-                  descriptionValue=""
-                  descriptionInput={false}
-                  deleteButton={false}
-                  onSubmit={() => addTag()}
-                  onClose={closeModal}
-                  onDelete={() => {}}
-                />
-              </div>
-            </div>
-          )}
-        </div>
+          </div>
+        )}
       </div>
     </div>
   );
